requirements.txt
# Core ML & DL libraries
tensorflow>=2.10.0
keras>=2.10.0

# Computer Vision
opencv-python>=4.7.0
opencv-contrib-python>=4.7.0

# Data processing
numpy>=1.23.0
pandas>=1.5.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Machine Learning utilities
scikit-learn>=1.2.0

# Configuration
pyyaml>=6.0

# Progress bars
tqdm>=4.64.0

# Image augmentation
albumentations>=1.3.0

# Model visualization
pydot>=1.4.2
graphviz>=0.20

# Jupyter support
jupyter>=1.0.0
ipykernel>=6.20.0
notebook>=6.5.0

# Utilities
pillow>=9.3.0
